+++ R/zzz_test_helper.R
@@ -78,3 +78,34 @@ get_df_portf_test <- function() {
 
 
 }
+
+#' @title simulate test data events
+#' @description generates multi-event data using sim_test_data_study()
+#' @param n_pat integer, number of patients, Default: 100
+#' @param n_sites integer, number of sites, Default: 5
+#' @param event_per_visit_mean mean event per visit per patient, Default: 0.5
+#' @param event_rates vector with visit-specific event rates, Default: Null
+#' @param event_names vector, contains the event names, default = "event"
+#' @return tibble with columns site_id, patient_id, is_ur, max_visit_mean,
+#'   max_visit_sd, visit, and event data (events_per_visit_mean and n_events)
+#' @keywords internal
+sim_test_data_events <- function(
+    n_pat = 100,
+    n_sites = 5,
+    event_per_visit_mean = 0.5,
+    event_rates = c(NULL),
+    event_names = list("event")) {
+
+  set.seed(1)
+  df_visit3 <- sim_test_data_study(n_pat = 100, n_sites = n_sites,
+                                   frac_site_with_ur = 0.4, ur_rate = 0.6,
+                                   event_per_visit_mean = event_per_visit_mean, event_names = event_names)
+  df_visit3$study_id <- "A"
+  set.seed(2)
+  df_visit4 <- sim_test_data_study(n_pat = 100, n_sites = n_sites,
+                                   frac_site_with_ur = 0.2, ur_rate = 0.1,
+                                   event_per_visit_mean = event_per_visit_mean, event_names = event_names)
+  df_visit4$study_id <- "B"
+  df_visit_events_test <- dplyr::bind_rows(df_visit3, df_visit4)
+  return(df_visit_events_test)
+}