+++ NEWS.md
@@ -3,6 +3,7 @@
 - eval_sites will combine over- and under-reporting probability into one column prob
 - eval_sites with inframe method will calculate delta events
 - new simaerep defaults inframe = TRUE, mult_corr = FALSE, visit_med75 = FALSE, under_only = FALSE
+- retire sim_ur_scenarios and get_portf_perf, portfolio performance will be calculated by removing events directly from the data with sim_ur
 
 # simaerep 0.7.0
 - event_names argument, allow to calculate multiple reporting probabilities at once https://github.com/openpharma/simaerep/issues/73